web-scraper-agent/src/requirements.txt
openai
tiktoken
python-dotenv
playwright
playwright-stealth
browser-use
scrapy
scrapy-playwright
selectolax
beautifulsoup4
lxml
requests
httpx
tenacity
fake-useragent
loguru
jinja2
pandas
aiofiles
pytest
python-wappalyzer